prevent syncing branch to wrong environment
I hit a scenario where Claude compacted the message history and then
created a new environment. It tried to sync my local changes to
'recover' the progress, which is unfortunate since it's only meant to do
that when the user explicitly requests it.

So, this adds a layer of defensiveness: now when you enable tracking for
a branch, we also track which environment the branch is tracking, and
error early if there's a mismatch.

Signed-off-by: Alex Suraci <suraci.alex@gmail.com>